pyproject.toml
[project]
name = "tophatc-clan-bot"
version = "1.0.0"
description = "Discord bot for TophatC Clan with Roblox integration"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "discord.py>=2.3.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["commands"]
include = [
    "bot.py",
    "database.py",
    "config.py",
    "roblox_api.py",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

